refactor: remove ProjectFormEdit.tsx
